+++ Assets/Scripts/Core/PlayersRoster.cs
@@ -1,0 +1,11 @@
+namespace Core
+{
+    public static class PlayersRoster
+    {
+        public enum PlayersList
+        {
+            Player1 = 1,
+            Player2 = 2
+        }
+    }
+}
