--- Assets/Scripts/PlayersRoster.cs
@@ -1,8 +1,0 @@
-public static class PlayersRoster
-{
-    public enum PlayersList
-    {
-        Player1 = 1,
-        Player2 = 2
-    }
-}
