+++ Assets/Scripts/View/MainMenuView.cs
@@ -1,4 +1,5 @@
 using Common;
+using Core;
 using Elements;
 using UnityEngine;
 using UnityEngine.SceneManagement;
